finaly

+++ Infrastructure.EfCore/Migrations/OAContextModelSnapshot.cs
@@ -84,9 +84,9 @@ namespace Infrastructure.EfCore.Migrations
 
             modelBuilder.Entity("Damain.Comment.Comment", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<int>("ArticleId")
